Add equisampling method to dataset generator as an option
Equisampling is essentially selecting images_per_person number of
images from each class to ensure each batch can generate a proper
number of triplets during training.
When this option is disabled, the presence of triplets in a batch
depends on luck. This makes each step faster (since dataset ops can
be completely converted to a graph), but convergence is much slower.