Remove cosine decay warm restarts policy
This rolls back the required Tensorflow version to 2.4.1.

Tensorflow 2.5.0 has a bug which prevents EfficientNets from being
loaded when mixed precision is active. The bug was fixed in the nightly
version, but nightly builds aren't supported by Tensorflow Addons.
Thus we have to wait for the official build to release, and then this
will be brought back.