Add new method for dataset generation to reduce memory footprint
This works by batching the dataset and then loading the images, instead
of loading the images first and then batching. This should ideally
reduce the memory consumed, as AUTOTUNE will determine how many batches
are pre-fetched. This also fixes the concern of batches not being
shuffled properly, which means that the neural network was likely to see
batches in similar order in each epoch.

The old function for creating the model was deprecated and has been
removed. Model creation will now be done using create_neural_network_v2
only.