Fix ESAM and SAM bugs
The issue was not with the optimizer; it was because of the conditional
statements used in ESAM. There do not seem to be any more bugs for now,
but the memory usage of ESAM needs to be examined. Currently, I am using
a hack where I set the gradient to 0; however, the gradient update
computation is still going to occur, just with no update. This needs to
be resolved. There are no noticeable performance gains compared to SAM
when testing with MNIST as well, which is another issue to look at.